GEP 6953 Validation Hotlist

* Hotlist name displayed on table.
Hotlist filter work.
Due to Hotlist work join is implemented and service calling is reverted.

* State name column work.

* Alpr server port changed to 8098

* Merge branch 'feature/development' into GEP_6781_CapturedPlates_Hotlist_Filter

* Merging issues solved.

* Do not delete hotlist if any number plate is associated with it.
Delete multiple hotlist support added.

* merged from developement.

* Fixed issues after merging development.

* Some code review changes.

* Handled invalid parameter exception.


Approved-by: Mehmood Usman

+++ Source/Crossbones.ALPR.Models/Items/HotListDTO.cs
@@ -13,7 +13,7 @@
 
         public string RulesExpression { get; set; }
 
-        public short AlertPriority { get; set; }
+        public int AlertPriority { get; set; }
 
         public DateTime CreatedOn { get; set; }
